ArticleLayout Component Refactoring
